pyproject.toml
[tool.poetry]
name = "autozzz"
version = "0.1.0"
description = ""
authors = ["Rain <iheyu22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyautogui = "^0.9.54"
pillow = "^10.4.0"
keyboard = "^0.13.5"
pydirectinput = "^1.0.4"
oathtool = "^2.3.1"
pyperclip = "^1.9.0"
ttkbootstrap = "^1.10.1"
loguru = "^0.7.2"
pygetwindow = "^0.0.9"
tesseract = "^0.1.3"
opencv-python = "^4.10.0.84"
pytesseract = "^0.3.10"
cnocr = "^2.3.0.3"
numpy = "<2"
onnxruntime = "^1.18.1"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
deptry = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
